Updates to gamestate handling, debug, full-screen.
Debug files are now loaded only if debug is on.
Changes to how gamestates are changed. The actual change now occurs at the top of the gameLoop.
Can now run the game full-screen by double-clicking the game canvas.
Updates to debug.
Added feature to create a general timer that operates similar to how the endDelay timers worked.
Started a concept for a basic grid-aligned "sprite" engine.
Refactor of gs_title1.